pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "particleplayground"
dynamic = ["version"]
authors = [
    {name = "James David Clarke", email = "james@jamesdavidclarke.com"},
]
description = "A Particle Playground game similar to Powder Toy"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
    "pygame",
    "appdirs",
    "python-dotenv",
    "tomli-w",
    "pillow"
]

[project.scripts]
particleplayground = "src.game:main"

[project.optional-dependencies]
dev = [
    "sphinx",
    "black",
    "watchdog"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/__init__.py"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''